feat: Move ChangeLogRequestParams to parent package [DHIS2-16964]
