+++ dhis-2/dhis-web-api/src/main/java/org/hisp/dhis/webapi/controller/tracker/export/trackedentity/ChangeLogRequestParamsMapper.java
@@ -36,6 +36,7 @@ import org.hisp.dhis.feedback.BadRequestException;
 import org.hisp.dhis.tracker.export.trackedentity.TrackedEntityChangeLogOperationParams;
 import org.hisp.dhis.tracker.export.trackedentity.TrackedEntityChangeLogOperationParams.TrackedEntityChangeLogOperationParamsBuilder;
 import org.hisp.dhis.webapi.controller.event.webrequest.OrderCriteria;
+import org.hisp.dhis.webapi.controller.tracker.export.ChangeLogRequestParams;
 
 class ChangeLogRequestParamsMapper {
   private ChangeLogRequestParamsMapper() {
